requirements.txt
torch
torchvision
torchaudio
transformers
datasets
PyMuPDF
pdf2image
pytesseract
opencv-python
Pillow
pyqt6
sentence-transformers